+++ shuriken/lib/parser/Dex/dex_classes.cpp
@@ -39,7 +39,7 @@ void ClassDataItem::parse_class_data_item(common::ShurikenStream& stream,
         access_flags = stream.read_uleb128();
         //! create the static field
         static_fields[static_field] =
-                std::make_unique<EncodedField>(fields.get_field_by_id(static_field),
+                std::make_unique<EncodedField>(fields.get_field_by_id(static_cast<uint32_t>(static_field)),
                                                                           static_cast<shuriken::dex::TYPES::access_flags>(access_flags));
 
     }
@@ -48,7 +48,7 @@ void ClassDataItem::parse_class_data_item(common::ShurikenStream& stream,
         instance_field += stream.read_uleb128();
         access_flags = stream.read_uleb128();
         instance_fields[instance_field] =
-                std::make_unique<EncodedField>(fields.get_field_by_id(instance_field),
+                std::make_unique<EncodedField>(fields.get_field_by_id(static_cast<uint32_t>(instance_field)),
                                                static_cast<shuriken::dex::TYPES::access_flags>(access_flags));
     }
 
@@ -58,7 +58,7 @@ void ClassDataItem::parse_class_data_item(common::ShurikenStream& stream,
         // for the code item
         code_offset = stream.read_uleb128();
         direct_methods[direct_method] =
-                std::make_unique<EncodedMethod>(methods.get_method_by_id(direct_method),
+                std::make_unique<EncodedMethod>(methods.get_method_by_id(static_cast<uint32_t>(direct_method)),
                                                 static_cast<shuriken::dex::TYPES::access_flags>(access_flags));
         direct_methods[direct_method]->parse_encoded_method(stream, code_offset, types);
     }
@@ -67,7 +67,7 @@ void ClassDataItem::parse_class_data_item(common::ShurikenStream& stream,
         virtual_method += stream.read_uleb128();
         access_flags = stream.read_uleb128();
         code_offset = stream.read_uleb128();
-        virtual_methods[virtual_method] = std::make_unique<EncodedMethod>(methods.get_method_by_id(virtual_method),
+        virtual_methods[virtual_method] = std::make_unique<EncodedMethod>(methods.get_method_by_id(static_cast<uint32_t>(virtual_method)),
                                                                           static_cast<shuriken::dex::TYPES::access_flags>(access_flags));
         virtual_methods[virtual_method]->parse_encoded_method(stream, code_offset, types);
     }
